test/test_media.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-


'''
Copyright 2010 Beech Horn

This file is part of lesscss-python.

lesscss-python is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

lesscss-python is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with lesscss-python.  If not, see <http://www.gnu.org/licenses/>.
'''


import unittest
from lesscss.media import parse_media


class TestMedia(unittest.TestCase):
    def test_multi(self):
        self.assertEqual(parse_media('@media screen { }').media, ['screen'])
        
    def test_none(self):
        self.assertRaises(ValueError, parse_media, '@media { }')
        
    def test_single(self):
        self.assertEqual(parse_media('@media screen, print { }').media,
                         ['screen', 'print'])


def suite():
    test_cases = (TestMedia,)
    
    suite = unittest.TestSuite()
    
    for tests in map(unittest.TestLoader().loadTestsFromTestCase, test_cases):
        suite.addTests(tests)

    return suite


if __name__ == '__main__':
    unittest.TextTestRunner(verbosity=2).run(suite())